Move location saving out of distance calculator
Not following the logic properly, I put the commands to save the current
location in Redis in the distance calculator, right after the distance
between last (currently in Redis) and current (freshly obtained from
GET) is calculated.  However, the distance calculator function is only
reached if there is both a current AND last position (to calculate a
distance between), meaning the code to save the current location as the
last location is never even reached.

I've moved the location saving out into after the branch the distance
calculator function is run, so that it is run in all cases when we
receive a location.  Fixes #1.