Round to 3 significant figures
This stops the weirdly overly precise distance figure from overflowing
on the screen, which fixes #2.  Also, for triple digit km values, I cast
the float to integer to get rid of the trailing '.0' that floats have,
because at triple digits all the significant figures are to the left of
the decimal, so that decimal just gives a false sense of accuracy.