Write out text using local URLs instead of yotext
The link sent back to the Yo-er used to point to yotext.co with text in
get parameters which would then be displayed nicely in "Yo style" but it
seems that service doesn't exist any more.

Now YouMoved itself has a new path "/show" which mimics the behaviour of
yotext, except that so far it's completely unstyled.  At least now we
get text again, instead of yotext's 404 page.  Fixes #4.